整合 客户信息中显示 和维护 客户分类

+++ src/main/resources/static/assets/js/customer/edit.js
@@ -1,41 +1,42 @@
 $().ready(function () {
-
     // 显示此客户对应的合同列表
-    contract_table = $("#contract_table").DataTable({
-        "width": "100%",
-        "autoWidth": true,
-        "ajax": {
-            url: contextPath + "contracts/"+ $("#id").val() +"/getDataWithoutPaged.json",
-            error: function (jqXHR, textStatus, errorThrown) {
-                var responseText = JSON.parse(jqXHR.responseText);
-                showMessage("error", "错误", responseText.data[0].errorMessage);
+    if ($("#contract_table").length > 0) {
+        contract_table = $("#contract_table").DataTable({
+            "width": "100%",
+            "autoWidth": true,
+            "ajax": {
+                url: contextPath + "contracts/" + $("#id").val() + "/getDataWithoutPaged.json",
+                error: function (jqXHR, textStatus, errorThrown) {
+                    var responseText = JSON.parse(jqXHR.responseText);
+                    showMessage("error", "错误", responseText.data[0].errorMessage);
+                },
             },
-        },
-        "language": {
-            "url": contextPath + "assets/js/lib/DataTables-1.10.16/chinese.lang.json"
-        },
-        "columns": [{
-            "data": "name"
-        }, {
-            "data": "description"
-        }, {
-            "data": "amount"
-        }, {
-            "data": "beginDate"
-        }, {
-            "data": "endDate"
-        }],
-        "columnDefs": [
-            {
-                "orderable": false,
-                "targets": [5],
-                "render": function (data, type, row, meta) {
-                    // return '<button type="button" class="btn btn-outline-primary"  onclick="fun_edit(\'' + row.id + '\')">编辑</button>&nbsp;&nbsp;<button class="btn btn-outline-danger btn-sm" type="button"  onclick="fun_delete(\'' + row.id + '\')">删除</button>'
-                    return '<a href="javascript:fun_contract_edit(\'' + row.id + '\')" class="m-portlet__nav-link btn m-btn m-btn--hover-accent m-btn--icon m-btn--icon-only m-btn--pill" title="编辑"><i class="la la-edit"></i></a>' +
-                        '<a href="javascript:fun_contract_delete(\'' + row.id + '\')" class="m-portlet__nav-link btn m-btn m-btn--hover-danger m-btn--icon m-btn--icon-only m-btn--pill" title=" 删除"><i class="la la-trash"></i></a>'
-                }
+            "language": {
+                "url": contextPath + "assets/js/lib/DataTables-1.10.16/chinese.lang.json"
+            },
+            "columns": [{
+                "data": "name"
+            }, {
+                "data": "description"
+            }, {
+                "data": "amount"
+            }, {
+                "data": "beginDate"
+            }, {
+                "data": "endDate"
             }],
-    });
+            "columnDefs": [
+                {
+                    "orderable": false,
+                    "targets": [5],
+                    "render": function (data, type, row, meta) {
+                        // return '<button type="button" class="btn btn-outline-primary"  onclick="fun_edit(\'' + row.id + '\')">编辑</button>&nbsp;&nbsp;<button class="btn btn-outline-danger btn-sm" type="button"  onclick="fun_delete(\'' + row.id + '\')">删除</button>'
+                        return '<a href="javascript:fun_contract_edit(\'' + row.id + '\')" class="m-portlet__nav-link btn m-btn m-btn--hover-accent m-btn--icon m-btn--icon-only m-btn--pill" title="编辑"><i class="la la-edit"></i></a>' +
+                            '<a href="javascript:fun_contract_delete(\'' + row.id + '\')" class="m-portlet__nav-link btn m-btn m-btn--hover-danger m-btn--icon m-btn--icon-only m-btn--pill" title=" 删除"><i class="la la-trash"></i></a>'
+                    }
+                }],
+        });
+    }
 });
 
 
@@ -61,6 +62,7 @@ function fun_submit() {
             window.location = contextPath + "customers/index";
         },
         error: function (data, textStatus, jqXHR) {
+            Logger.debug(jqXHR);
             var responseText = JSON.parse(jqXHR.responseText);
             mApp.unblock();
             showMessage("error", "错误", responseText.data[0].errorMessage);
